backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
alembic>=1.13.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0
openpyxl>=3.1.0
joblib>=1.3.0
python-dotenv>=1.0.0
